requirements.txt
Flask==0.12.2
RPi.GPIO==0.6.3
apscheduler==3.5.1
